application/requirements.txt
application==1.0
Flask==2.0.1
mysql_connector_repackaged==0.3.1
mysql.connector==2.2.9
